requirements.txt
pytorch-gpu
cudatoolkit
numpy
ase
pymol-open-source
rdkit
bs4
mechanize
openbabel
packmol
